feat(rdb): extract the CVE ID set with CPE
