+++ db/rdb.go
@@ -484,8 +484,6 @@ func (r *RDBDriver) GetCveIDsByCpeURI(uri string) (nvdCveIDs []string, jvnCveIDs
 	}
 
 	// The cpes table is de-normalized. So the `First` is correct, don't be mislead.
-	nvdCveIDs = []string{}
-	jvnCveIDs = []string{}
 	uniqCveIDs := map[string]bool{}
 	for _, cpe := range nvdCpes {
 		nvd := models.NvdJSON{}
@@ -494,7 +492,6 @@ func (r *RDBDriver) GetCveIDsByCpeURI(uri string) (nvdCveIDs []string, jvnCveIDs
 			return nil, nil, err
 		}
 		if _, ok := uniqCveIDs[nvd.CveID]; !ok {
-			nvdCveIDs = append(nvdCveIDs, nvd.CveID)
 			uniqCveIDs[nvd.CveID] = true
 		}
 	}
@@ -506,8 +503,27 @@ func (r *RDBDriver) GetCveIDsByCpeURI(uri string) (nvdCveIDs []string, jvnCveIDs
 			return nil, nil, err
 		}
 		if _, ok := uniqCveIDs[jvn.CveID]; !ok {
-			jvnCveIDs = append(jvnCveIDs, jvn.CveID)
 			uniqCveIDs[jvn.CveID] = true
+		}
+	}
+
+	nvdCveIDs = []string{}
+	jvnCveIDs = []string{}
+	for cveID := range uniqCveIDs {
+		d, err := r.Get(cveID)
+		if err != nil {
+			return nil, nil, err
+		}
+		nvdMatch, jvnMatch, err := matchCpe(uri, d)
+		if err != nil {
+			log.Warnf("Failed to compare the version:%s %s %#v",
+				err, uri, d)
+			continue
+		}
+		if nvdMatch {
+			nvdCveIDs = append(nvdCveIDs, d.CveID)
+		} else if jvnMatch {
+			jvnCveIDs = append(jvnCveIDs, d.CveID)
 		}
 	}
