+++ db/redis.go
@@ -266,65 +266,6 @@ func (r *RedisDriver) GetByCpeURI(uri string) ([]models.CveDetail, error) {
 	return details, nil
 }
 
-func matchCpe(uri string, cve *models.CveDetail) (nvdMatch, jvnMatch bool, err error) {
-	nvdCpes := []models.Cpe{}
-	if cve.NvdJSON != nil {
-		nvdCpes = append(nvdCpes, cve.NvdJSON.Cpes...)
-	}
-	for _, cpe := range nvdCpes {
-		match, err := match(uri, cpe)
-		if err != nil {
-			log.Debugf("Failed to match: %s", err)
-
-			// Try to exact match by vendor, product and version if the version in CPE is not a semVer style.
-			ok, err := matchExactByAffects(uri, cve.NvdJSON.Affects)
-			if err != nil {
-				return false, false, err
-			}
-			if ok {
-				return true, false, nil
-			}
-			continue
-		} else if match {
-			return true, false, nil
-		}
-	}
-
-	// CPE that exists only in JVN is also detected.
-	// There is a possibility of false positives since the JVN does not contain version information.
-	if cve.Jvn == nil {
-		return false, false, nil
-	}
-
-	for _, cpe := range cve.Jvn.Cpes {
-		// If NVD has data of the same `type`, `verndor`, and `product`, NVD is used in priority.
-		// Because NVD has version information, but JVN does not.
-		found := false
-		for _, nvdCpe := range nvdCpes {
-			ok, err := isSamePartVendorProduct(nvdCpe.URI, cpe.URI)
-			if err != nil {
-				continue
-			}
-			if ok {
-				found = true
-				break
-			}
-		}
-		if found {
-			continue
-		}
-
-		ok, err := isSamePartVendorProduct(uri, cpe.URI)
-		if err != nil {
-			continue
-		}
-		if ok {
-			return false, true, nil
-		}
-	}
-	return false, false, nil
-}
-
 // InsertJvn insert items fetched from JVN.
 func (r *RedisDriver) InsertJvn(cves []models.CveDetail) error {
 	ctx := context.Background()
