+++ db/redis_test.go
@@ -1,187 +1,1 @@
 package db
-
-import (
-	"testing"
-
-	"github.com/kotakanbe/go-cve-dictionary/models"
-)
-
-func Test_matchCpe(t *testing.T) {
-	nvdRangeCpe := models.Cpe{
-		NvdJSONID: 1,
-		CpeBase: models.CpeBase{
-			URI: "cpe:/o:qualcomm:qcs605_firmware",
-			CpeWFN: models.CpeWFN{
-				Part:    "o",
-				Vendor:  "qualcomm",
-				Product: "qcs605_firmware",
-			},
-			VersionStartIncluding: "0.0.1",
-			VersionEndIncluding:   "2.0.0",
-		},
-	}
-	nvdNonSemVerCpe := models.Cpe{
-		NvdJSONID: 1,
-		CpeBase: models.CpeBase{
-			URI: "cpe:/o:qualcomm:qcs605_firmware",
-			CpeWFN: models.CpeWFN{
-				Part:    "o",
-				Vendor:  "qualcomm",
-				Product: "qcs605_firmware",
-			},
-			VersionStartIncluding: "hoge",
-			VersionEndIncluding:   "hoge",
-		},
-	}
-	jvnCpe := models.Cpe{
-		JvnID: 1,
-		CpeBase: models.CpeBase{
-			URI: "cpe:/o:qualcomm:qcs605_firmware",
-			CpeWFN: models.CpeWFN{
-				Part:    "o",
-				Vendor:  "qualcomm",
-				Product: "qcs605_firmware",
-			},
-		},
-	}
-	type args struct {
-		uri string
-		cve *models.CveDetail
-	}
-	tests := []struct {
-		name    string
-		args    args
-		nvdWant bool
-		jvnWant bool
-		wantErr bool
-	}{
-		{
-			name: "nvd range match",
-			args: args{
-				uri: "cpe:/o:qualcomm:qcs605_firmware:1.0.0",
-				cve: &models.CveDetail{
-					NvdJSON: &models.NvdJSON{
-						Cpes: []models.Cpe{nvdRangeCpe},
-					},
-				},
-			},
-			nvdWant: true,
-			wantErr: false,
-		},
-		{
-			name: "nvd range not match",
-			args: args{
-				uri: "cpe:/o:qualcomm:qcs605_firmware:3.0.0",
-				cve: &models.CveDetail{
-					NvdJSON: &models.NvdJSON{
-						Cpes: []models.Cpe{nvdRangeCpe},
-					},
-				},
-			},
-			nvdWant: false,
-			wantErr: false,
-		},
-		{
-			name: "nvd affect match",
-			args: args{
-				uri: "cpe:/o:qualcomm:qcs605_firmware:1.0.0",
-				cve: &models.CveDetail{
-					NvdJSON: &models.NvdJSON{
-						Cpes: []models.Cpe{nvdNonSemVerCpe},
-						Affects: []models.Affect{
-							{
-								Vendor:  "qualcomm",
-								Product: "qcs605_firmware",
-								Version: "1.0.0",
-							},
-						},
-					},
-				},
-			},
-			nvdWant: true,
-			wantErr: false,
-		},
-		{
-			name: "nvd affects not match",
-			args: args{
-				uri: "cpe:/o:qualcomm:qcs605_firmware:1.0.0",
-				cve: &models.CveDetail{
-					NvdJSON: &models.NvdJSON{
-						Cpes: []models.Cpe{nvdNonSemVerCpe},
-						Affects: []models.Affect{
-							{
-								Vendor:  "qualcomm",
-								Product: "qcs605_firmware",
-								Version: "1.0.1",
-							},
-						},
-					},
-				},
-			},
-			nvdWant: false,
-			wantErr: false,
-		},
-		{
-			name: "jvn match",
-			args: args{
-				uri: "cpe:/o:qualcomm:qcs605_firmware:1.0.0",
-				cve: &models.CveDetail{
-					Jvn: &models.Jvn{
-						Cpes: []models.Cpe{jvnCpe},
-					},
-				},
-			},
-			jvnWant: true,
-			wantErr: false,
-		},
-		{
-			name: "NVD has priority over JVN",
-			args: args{
-				uri: "cpe:/o:qualcomm:qcs605_firmware:3.0.0",
-				cve: &models.CveDetail{
-					NvdJSON: &models.NvdJSON{
-						Cpes: []models.Cpe{nvdRangeCpe},
-					},
-					Jvn: &models.Jvn{
-						Cpes: []models.Cpe{jvnCpe},
-					},
-				},
-			},
-			nvdWant: false,
-			jvnWant: false,
-			wantErr: false,
-		},
-		{
-			name: "NVD has priority over JVN",
-			args: args{
-				uri: "cpe:/o:qualcomm:qcs605_firmware:1.0.0",
-				cve: &models.CveDetail{
-					NvdJSON: &models.NvdJSON{
-						Cpes: []models.Cpe{nvdRangeCpe},
-					},
-					Jvn: &models.Jvn{
-						Cpes: []models.Cpe{jvnCpe},
-					},
-				},
-			},
-			nvdWant: true,
-			jvnWant: false,
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			nvdGot, jvnGot, err := matchCpe(tt.args.uri, tt.args.cve)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("matchCpe() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if nvdGot != tt.nvdWant {
-				t.Errorf("matchCpe() nvd = %v, want %v", nvdGot, tt.nvdWant)
-			}
-			if jvnGot != tt.jvnWant {
-				t.Errorf("matchCpe() jvn = %v, want %v", jvnGot, tt.jvnWant)
-			}
-		})
-	}
-}
